requirements_dev.txt
bentoml==1.0.25
#bentoml==1.0.10
joblib==1.2.0
pip-chill==1.0.1
torchvision==0.14.1
tqdm==4.64.1
wincertstore==0.2
dvc
mlflow
ipykernel
pandas
numpy
seaborn
pytest==7.1.3
tox==3.25.1
black==22.8.0
flake8==5.0.4
mypy==0.971
#-e .